% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{animateOff}
\alias{animateOff}
\title{animateOff}
\usage{
animateOff(active = FALSE, fileName, saveFrames = FALSE, frameRate = 25)
}
\description{
Second of two functions needed to create mpg from static images. This function should be run after running \code{animateOn()} and creating a series of plots. The \code{animateOff()} function should have the same shared arguments as \code{animateOn()}.
}

