% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{animateOn}
\alias{animateOn}
\title{animateOn}
\usage{
animateOn(active = FALSE, fileName, saveFrames = FALSE, width = 480,
  height = 480, units = "px", pointsize = 12, quality = 75,
  bg = "white", res = NA)
}
\arguments{
\item{active}{allows function to be selectively de-activated using a variable.}

\item{fileName}{the filename of input files (without extension or number). The same name is used for the final video file.}

\item{saveFrames}{if FALSE then images are saved to a temporary location before being deleted in \code{animateOff()}.}

\item{width}{width of video.}

\item{height}{height of video.}

\item{units}{units of jpg files that make up video (see ?jpeg).}

\item{pointsize}{pointsize of jpg files that make up video (see ?jpg).}

\item{quality}{quality of jpg files that make up video (see ?jpg).}

\item{bg}{background of jpg files that make up video (see ?jpg).}

\item{res}{resolution of jpg files that make up video (see ?jpg).}
}
\description{
First of two functions needed to create mpg from static images. Run this function, then run the code needed to create a sequence of plots, then run \code{animateOff()}. Individual plots will be saved as jpg with a four-character number on the end (for example myFile0012.jpg), and can either be saved in the local directory or in a temporary directory. The \code{animateOff()} function should have the same shared arguments as \code{animateOn()} (for example the same fileName).
\cr\cr
Note that this function requires ImageMagick to be installed, and has only been tested on Windows. If running for the first time you will likely need to add ImageMagick to the path using something like the following:
\code{Sys.setenv(PATH=paste(Sys.getenv("PATH"),"C:\\Program Files\\ImageMagick-6.9.3-Q16",sep=";"))}
}
