% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convertRadix}
\alias{convertRadix}
\title{convertRadix}
\usage{
convertRadix(x, base = 2, fixlength = NA)
}
\arguments{
\item{x}{number to convert.}

\item{base}{radix to convert to.}

\item{fixlength}{if used, always output at least this many digits.}
}
\description{
Convert decimal number system to any other radix (base).
}

