% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cubicSpline}
\alias{cubicSpline}
\title{cubeSpline}
\usage{
cubicSpline(x, y, g, interpoints = 10)
}
\description{
Calculate cubic spline passing through the points (x,y) with known gradients g at these points. The parameter 'interpoints' describes the number of points in each segment.
}

