% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dBDI}
\alias{dBDI}
\title{dBDI}
\usage{
dBDI(x, birthRate, deathRate, immigrationRate, log = FALSE)
}
\arguments{
\item{x}{number of individuals or genes (i.e. value of random variable).}

\item{birthRate}{rate of birth.}

\item{deathRate}{rate of death.}

\item{immigrationRate}{rate of immigration.}
}
\description{
Probability mass function for the birth-death-immigration model. Note that this is simply a particular version of the negative-binomial likelihood. When applied to gene families birthRate=rate of gene duplication, deathRate=rate of gene deletion, immigrationRate=rate of de novo gene creation.
}
