% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{densityPlot}
\alias{densityPlot}
\title{densityPlot}
\usage{
densityPlot(x, xmin = NULL, xmax = NULL, ymin = 0, ymax = NULL,
  xlab = "", ylab = "density", main = "", meanOn = TRUE,
  medianOn = TRUE, quantileOn = TRUE, boxOn = TRUE, boxSize = 0.8,
  meanTextOn = TRUE, medianTextOn = TRUE, quantileTextOn = TRUE,
  boxSigFig = 3)
}
\description{
Produce kernel density plot of data x, with mean, median and 95% limits shown.
}

