% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{errorBars}
\alias{errorBars}
\title{errorBars}
\usage{
errorBars(y1, y2, x = 1:length(y1), se = FALSE, width = 1, col = 1,
  lty = 1, lwd = 1)
}
\description{
Produce error bars at given positions. When se=FALSE, input should be y1=lower limit and y2=upper limit. When se=TRUE, input should be y1=mean and y2=standard error, in which case error bars represent two standard errors either side of the mean. Can handle vector inputs as well as scalars.
}

