% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{faster}
\alias{faster}
\title{faster}
\usage{
faster(..., col = NULL, breaks = NULL, nlevel = 64, layout_mat = NULL,
  horizontal = FALSE, legend.shrink = 0.9, legend.width = 1.2,
  legend.mar = ifelse(horizontal, 3.1, 5.1), legend.lab = NULL,
  legend.line = 2, bigplot = NULL, smallplot = NULL, lab.breaks = NULL,
  axis.args = NULL, legend.args = NULL, legend.cex = 1)
}
\description{
Custom version of raster plot, that is lightweight and doesn't suffer from problems relating to layout. Ordinary raster produces strange results when used as part of more complex layouts due to its method of setting and re-setting par(). This is avoided here by passing in an optional layout matrix that corresponds to the current layout.
}
