% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gcDist}
\alias{gcDist}
\title{great circle distance}
\usage{
gcDist(origin_lat, origin_lon, dest_lat, dest_lon)
}
\arguments{
\item{origin_lat}{scalar latitude of origin in degrees}

\item{origin_lon}{scalar longitude of origin in degrees}

\item{origin_lat}{vector latitude of destination in degrees}

\item{origin_lon}{vector longitude of destination in degrees}
}
\description{
Calculates great circle distance (km) between an origin and one or more destination points.
}
