% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getListDepth}
\alias{getListDepth}
\title{get list depth}
\usage{
getListDepth(x)
}
\arguments{
\item{x}{list or nested list to assess depth}
}
\description{
Get list depth up to 5 levels of nesting. Non-lists return a value 0.
}
