% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{grad}
\alias{grad}
\title{grad}
\usage{
grad(x = 1:length(y), y)
}
\description{
Numerically calculate gradient at each point (x,y). The gradient at a point is calculated as the mean of the two gradients in either direction from the point, apart from at the ends of the vector where there is only a single value to consider.
}
