% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{imageFix}
\alias{imageFix}
\title{imageFix}
\usage{
imageFix(z, x = 1:nrow(z), y = 1:ncol(z), orientation = 1, ...)
}
\description{
Produces an image plot, but has option for changing orientation. Values of \code{orientation} from 1 to 4 rotate the image clockwise.
}
