% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{incrementRestrictedGrowth}
\alias{incrementRestrictedGrowth}
\title{incrementRestrictedGrowth}
\usage{
incrementRestrictedGrowth(group, K)
}
\description{
Find next group that satisfies restricted growth function (max value K). Allows user to eplore all unique partitions by iteratively updating the grouping. Returns 0 when no more groups.
}

