% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{latexTable}
\alias{latexTable}
\title{latexTable}
\usage{
latexTable(x, digits = 3, file = "~/Desktop/tester.txt")
}
\description{
Convert data frame x to LaTeX format with set decimal places, and write output to file.
}

