% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loadPackage}
\alias{loadPackage}
\title{loadPackage}
\usage{
loadPackage(string)
}
\description{
Single command for installing (if not already installed) and loading a package. For example, the \code{loadPackage('devtools')} command would run the following code block:
\preformatted{
if (!"devtools"\%in\%rownames(installed.packages())) {
	install.packages('devtools')
}
require("devtools")
}
}
