% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{logDescriptive}
\alias{logDescriptive}
\title{logDescriptive}
\usage{
logDescriptive(Y)
}
\description{
Calculates descriptive stats on extremely large or extremely small values. Input vector of values in log-space. Output sample mean and variance, also in log-space.
}
