% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{logHarmonicMean}
\alias{logHarmonicMean}
\title{logHarmonicMean}
\usage{
logHarmonicMean(z)
}
\description{
Calculates harmonic mean of values z, where values are in log space. Accounts for underflow and returns value of log harmonic mean.
}

