% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{margin}
\alias{margin}
\title{change margins}
\usage{
margin(level)
}
\arguments{
\item{level}{choose from a number of settings: 1) default margins c(5.1,4.1,4.1,2.1), 2) narrow margins c(1,3,3,1).}
}
\description{
Changes margins to one of a number of defaults
}
