% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{matrixSmooth}
\alias{matrixSmooth}
\title{matrixSmooth}
\usage{
matrixSmooth(M, reps = 1)
}
\description{
Interpolate rows and columns of a matrix alternately using spline. reps is the number of times to double the matrix size.
}
