% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MCMCPlot2}
\alias{MCMCPlot2}
\title{MCMCPlot2}
\usage{
MCMCPlot2(x, h_cells = 100, v_cells = 100, xmin = 1, xmax = length(x),
  ymin = min(x, na.rm = T), ymax = max(x, na.rm = T), xlab = "iteration",
  main = "", ylab = main)
}
\description{
Similar to \code{MCMCPlot()}, but aimed at very large numbers of samples where ordinary plotting would not be feasible. Works by binning points before plotting. Can only handle a single chain (cannot plot x vs. y).
}

