% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{merge.SpatialPolygonsDataFrame}
\alias{merge.SpatialPolygonsDataFrame}
\title{merge.SpatialPolygonsDataFrame}
\usage{
\method{merge}{SpatialPolygonsDataFrame}(shp, df)
}
\description{
Reads in a shape file and data frame to be merged with the data in this shapefile. Merges while preserving the order of objects (ordinary merge operation causes polygons to become disassociated with data).
}
