% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{minSpanTree}
\alias{minSpanTree}
\title{minSpanTree}
\usage{
minSpanTree(data)
}
\description{
Use Prim's algorithm to calculate minimum spanning tree. The 'data' argument must be a matrix or data frame with observations in rows and dimensions in columns. Distances are calculated as Euclidean distance over all the dimensions of the data. Returns multiple objects; 1) a data frame of nodePairs giving all pairwise links that make up the tree in the order that they were added, 2) an edgeList giving the nodes attached to each node, 3) edgeNum giving the number of edges associated with each node.
\cr\cr
Although this function may not be as fast as some alternatives (not tested), it has the advantage that distances are only calculated as needed, meaning there is no need to read in a huge matrix of pairwise distances.
}
