% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{perlin_noise}
\alias{perlin_noise}
\title{perlin_noise}
\usage{
perlin_noise(outRows = 100, outCols = 100, levelsX = 10, levelsY = 10)
}
\arguments{
\item{outRows}{rows in output matrix.}

\item{outCols}{columns in output matrix.}

\item{levelsX}{bumpyness in x dimension.}

\item{levelsY}{bumpyness in y dimension.}
}
\description{
Generates Perlin noise of any scale in, a matrix of any size. Raw code is copied from internet.
}

