% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pieCharts}
\alias{pieCharts}
\title{pieCharts}
\usage{
pieCharts(x, y, proportions, radius = 0.2, lwd = 1, border_col = 1,
  seg_col = bobRainbow(), smoothness = 50, x_stretch = 1)
}
\description{
Adds pie charts to an existing plot. Proportions do not need to sum to one.
}

