% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotOn}
\alias{plotOn}
\title{plotOn}
\usage{
plotOn(active = FALSE, fileroot = "", filestem1 = "", filestem2 = "",
  fileindex = "", type = "pdf", width = 6, height = 6, res = 300)
}
\description{
start optional save-to-file function
}

