% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{RM1_stochastic_hybrid}
\alias{RM1_stochastic_hybrid}
\title{RM1_stochastic_hybrid}
\usage{
RM1_stochastic_hybrid(a = 0.3, p = 0.9, mu = NULL, u = 22, v = 10,
  r = 1/200, b = 1, c = 1, Eh_init = 0, Ih_init = 10, H = 100,
  Em_init = 0, Im_init = 0, M = 100, times = 0:100,
  maxIterations = 10000)
}
\arguments{
\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. mu = -log(p) unless specified.}

\item{u}{intrinsic incubation period. The number of days from infection to infectiousness in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to infectiousness in a mosquito host.}

\item{r}{daily recovery rate.}

\item{b}{probability a human becomes infected after being bitten by an infected mosquito.}

\item{c}{probability a mosquito becomes infected after biting an infected human.}

\item{Eh_init}{initial number of infected but not infectious humans.}

\item{Ih_init}{initial number of infectious humans.}

\item{H}{human population size.}

\item{Em_init}{initial number of infected but not yet infectious mosquitoes.}

\item{Im_init}{initial number of infectious mosquitoes.}

\item{M}{mosquito population size (number of adult female mosquitoes).}

\item{times}{vector of times at which output should be returned.}

\item{maxIterations}{exit if this number of iterations is reached.}
}
\description{
Draw from asynchronous stochastic Ross-Macdonald model using a compromise between a synchronous and an asynchronous algorithm. The basic algorithm is asynchronous (based on Gillespie's algorithm), but values are only stored and returned at discrete time points. The function still exits automatically at a defined maximum number of iterations.
}

