% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{RM2_deterministic}
\alias{RM2_deterministic}
\title{RM2_deterministic}
\usage{
RM2_deterministic(a = a, mu = 0.1, lambda = 0.2, u = 22, v = 10,
  r = 1/200, b = 1, c = 1, Eh_init = 0, Ih_init = 10, Em_init = 0,
  Im_init = 0, H = 100, M_init = 100, times = 0:100, Ktimes = c(0,
  100), Kvalues = c(100, 200))
}
\arguments{
\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on humans each day.}

\item{mu}{mosquito instantaneous  death rate.}

\item{lambda}{mosquito instantaneous birth rate.}

\item{u}{intrinsic incubation period. The number of days from infection to infectiousness in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to infectiousness in a mosquito host.}

\item{r}{daily recovery rate.}

\item{b}{probability a human becomes infected after being bitten by an infected mosquito.}

\item{c}{probability a mosquito becomes infected after biting an infected human.}

\item{Eh_init}{initial number of infected but not infectious humans.}

\item{Ih_init}{initial number of infectious humans.}

\item{Em_init}{initial number of infected but not yet infectious mosquitoes.}

\item{Im_init}{initial number of infectious mosquitoes.}

\item{H}{human population size.}

\item{M_init}{initial mosquito population size.}

\item{times}{vector of times at which output should be returned.}

\item{Ktimes}{vector of times at which carrying capacity is defined.}

\item{Kvalues}{vector of carrying capacities corresponding to \code{Ktimes}..}
}
\description{
Variation of RM1 model - in this version the mosquito population grows in logistic manner given growth and death rates and a carrying capacity. The carrying capacity can be varied at any given time point to allow for changing mosquito densities over time.
}

