% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{RM2_stochastic_sync}
\alias{RM2_stochastic_sync}
\title{RM2_stochastic_sync}
\usage{
RM2_stochastic_sync(a = 0.3, mu = 0.1, lambda = 0.2, u = 22, v = 10,
  r = 1/200, b = 1, c = 1, Eh_init = 0, Ih_init = 10, Em_init = 0,
  Im_init = 0, H = 100, M_init = 100, times = 0:100, Ktimes = c(0,
  100), Kvalues = c(100, 200))
}
\arguments{
\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on humans each day.}

\item{mu}{mosquito instantaneous  death rate.}

\item{lambda}{mosquito instantaneous birth rate.}

\item{u}{intrinsic incubation period. The number of days from infection to infectiousness in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to infectiousness in a mosquito host.}

\item{r}{daily recovery rate.}

\item{b}{probability a human becomes infected after being bitten by an infected mosquito.}

\item{c}{probability a mosquito becomes infected after biting an infected human.}

\item{Eh_init}{initial number of infected but not infectious humans.}

\item{Ih_init}{initial number of infectious humans.}

\item{Em_init}{initial number of infected but not yet infectious mosquitoes.}

\item{Im_init}{initial number of infectious mosquitoes.}

\item{H}{human population size.}

\item{M_init}{initial mosquito population size (number of adult female mosquitoes).}

\item{times}{vector of times at which output should be returned.}

\item{Ktimes}{vector of times at which carrying capacity is defined.}

\item{Kvalues}{vector of carrying capacities that come into action at \code{Ktimes}.}
}
\description{
Draw from a synchronous stochastic version of a particular Ross-Macdonald-style model (see \code{?RM2_deterministic} for details of the model). Return state of the system at known time points. Results of the synchronous method only match up with the asynchronous method when the time step is small relative to the rates that drive the system.
}

