% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rMultiMix}
\alias{rMultiMix}
\title{rMultiMix}
\usage{
rMultiMix(n, K, psi, nu, alpha, tau, pMissing = 0)
}
\arguments{
\item{n}{number of draws.}

\item{K}{number of mixture components. Set K=Inf to use infinite mixture model.}

\item{psi}{prior matrix for inverse Wishart prior on covariance. When psi is a scalar value the model is one-dimensional, and variances are effectively drawn from an inverse-gamma distribution with shape alpha=nu/2 and rate beta=psi/2.}

\item{nu}{degrees of freedom for inverse Wishart prior on covariance.}

\item{alpha}{Dirichlet prior parameter for mixture weights. Set alpha=Inf to use model of equal weights.}

\item{tau}{standard deviation of prior on individual components means (this prior is centred at 0).}

\item{pMissing}{proportion of data that is missing.}
}
\description{
Draw from finite or infinite (Dirichlet process) multivariate normal mixture model. Note that although correlations are allowed for each mixture component, the locations of components are drawn from a multivariate normal distribution with standard deviation tau in all directions (i.e. no correlation). Allows for generation of missing data by randomly eliminating a proportion of results.
}

