% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rowVars}
\alias{rowVars}
\title{rowVars}
\usage{
rowVars(mat, sampleVar = TRUE, mu = 0)
}
\description{
Calculate variance (sample or population) of rows of a matrix or data frame. If sampleVar==FALSE then calcualte relative to mean mu.
}

