% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{safeRead}
\alias{safeRead}
\title{safeRead}
\usage{
safeRead(fileName, delim = "\\t", useHeader = TRUE, report = TRUE,
  badCharacters = c("-1.#IND", "1.#INF", "-999"), replacements = c(0, 0, 0),
  overwrite = FALSE)
}
\description{
Reads in text file (tab-delimited by default) and replaces bad characters with replacements.
\cr\cr
WARNING - can overwrite original file with new values if \code{overwrite==TRUE}, although this is not the default.
}
