% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sampleVec}
\alias{sampleVec}
\title{sample always from vector}
\usage{
sampleVec(x, ...)
}
\arguments{
\item{x}{vector from which to sample}
}
\description{
Carries out sample(), but always assumes input is vector.
}
