% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SIR_delay_deterministic2}
\alias{SIR_delay_deterministic2}
\title{SIR_delay_deterministic}
\usage{
SIR_delay_deterministic2(beta = 0.5, dur_inf = 4, I_init = 10,
  R_init = 0, N = 1000, times = 0:100)
}
\arguments{
\item{beta}{contact rate.}

\item{dur_inf}{length of time in infectious state.}

\item{I_init}{initial number of infectious individuals.}

\item{R_init}{initial number of recovered (immune) individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}
}
\description{
Returns solution to deterministic SIR model in which individuals are infectious for a fixed amount of time and there is no natural birth and death. Solves delay differential equation using the \code{odin} package. A set number of infectious individuals \code{I_init} are assumed to be seeded at time \code{t=0}, and will recover simultaneously at time \code{t=dur_inf}.
}

