% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SIR_deterministic}
\alias{SIR_deterministic}
\title{SIR_deterministic}
\usage{
SIR_deterministic(beta = 1, r = 0.25, mu = 0.01, I_init = 100,
  R_init = 0, N = 1000, times = 0:100)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{mu}{natural death rate (same in all compartments). Also rate of new births into susceptible compartment.}

\item{I_init}{initial number of infectious individuals.}

\item{R_init}{initial number of recovered (immune) individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}
}
\description{
Returns solution to deterministic SIR model using the \code{odin} package.
}

