% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SIR_stochastic_async}
\alias{SIR_stochastic_async}
\title{SIR_stochastic_async}
\usage{
SIR_stochastic_async(beta = 1, r = 0.25, mu = 0.01, I_init = 100,
  R_init = 0, N = 1000, maxIterations = 10000)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{mu}{natural death rate (same in all compartments). Also rate of new births into susceptible compartment.}

\item{I_init}{initial number of infectious individuals.}

\item{R_init}{initial number of recovered (immune) individuals.}

\item{N}{total number of individuals in population.}

\item{maxIterations}{exit if this number of iterations is reached.}
}
\description{
Draw from asynchronous stochastic SIR model. Return state of the system at all time points at which any event occurs. Stop when maxIterations is reached. Note that natural deaths are exactly matched by births into the susceptible state in this formulation, meaning the population size stays constant at N.
}
