% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SIR_stochastic_sync}
\alias{SIR_stochastic_sync}
\title{SIR_stochastic_sync}
\usage{
SIR_stochastic_sync(beta = 1, r = 0.25, mu = 0.01, I_init = 100,
  R_init = 0, N = 1000, times = 0:100)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{mu}{natural death rate (same in all compartments). Also rate of new births into susceptible compartment.}

\item{I_init}{initial number of infectious individuals.}

\item{R_init}{initial number of recovered (immune) individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}
}
\description{
Draw from synchronous stochastic SIR model. Return state of the system at all time points at which an event occurs. Note that natural deaths are exactly matched by births into the susceptible state in this formulation, meaning the population size stays constant at N. Results of the synchronous method only match up with the asynchronous method when the time step is small relative to the rates that drive the system.
}

