% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SIS_deterministic}
\alias{SIS_deterministic}
\title{SIS_deterministic}
\usage{
SIS_deterministic(beta = 1, r = 0.25, I_init = 10, N = 1000,
  times = 0:100)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}
}
\description{
Returns solution to deterministic SIS model using the \code{odin} package.
}

