% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SIS_stochastic_hybrid}
\alias{SIS_stochastic_hybrid}
\title{SIS_stochastic_hybrid}
\usage{
SIS_stochastic_hybrid(beta = 1, r = 0.25, I_init = 100, N = 1000,
  times = 0:100, maxIterations = 10000)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}

\item{maxIterations}{exit if this number of iterations is reached.}
}
\description{
Draw from stochastic SIS model using a compromise between a synchronous and an asynchronous algorithm. The basic algorithm is asynchronous (based on Gillespie's algorithm), but values are only stored and returned at discrete time points. The function still exits automatically at a defined maximum number of iterations.
}
