% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SIS_stochastic_sync}
\alias{SIS_stochastic_sync}
\title{SIS_stochastic_sync}
\usage{
SIS_stochastic_sync(beta = 1, r = 0.25, I_init = 100, N = 1000,
  times = 0:100)
}
\arguments{
\item{beta}{contact rate.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}
}
\description{
Draw from synchronous stochastic SIS model. Return infectives at known time points. Note that results of the synchronous method only match up with the asynchronous method when the time step is small relative to the rates that drive the system.
}
