% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SLIR_deterministic}
\alias{SLIR_deterministic}
\title{SLIR_deterministic}
\usage{
SLIR_deterministic(beta = 0.5, dur_lag = 1, r = 0.25, I_init = 10,
  R_init = 0, N = 1000, times = 0:100)
}
\arguments{
\item{beta}{contact rate.}

\item{dur_lag}{length of time in incubation state.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{R_init}{initial number of recovered (immune) individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}
}
\description{
Returns solution to deterministic SLIR model, where L is an incubation (lag) stage of defined length. Solves delay differential equation using the \code{odin} package.
}

