% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SLIR_stochastic_async}
\alias{SLIR_stochastic_async}
\title{SLIR_stochastic_async}
\usage{
SLIR_stochastic_async(beta = 1, dur_lag = 1, r = 0.25, I_init = 100,
  R_init = 0, N = 1000, maxIterations = 10000)
}
\arguments{
\item{beta}{contact rate.}

\item{dur_lag}{length of time in incubation state.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{R_init}{initial number of recovered (immune) individuals.}

\item{N}{total number of individuals in population.}

\item{maxIterations}{exit if this number of iterations is reached.}
}
\description{
Draw from asynchronous stochastic SLIR model, where L is an incubation (lag) stage of defined length. Return state of the system at all time points at which any event occurs. Stop when maxIterations is reached.
}

