% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{SLIR_stochastic_hybrid}
\alias{SLIR_stochastic_hybrid}
\title{SLIR_stochastic_hybrid}
\usage{
SLIR_stochastic_hybrid(beta = 1, dur_lag = 1, r = 0.25, I_init = 100,
  R_init = 0, N = 1000, times = 0:100, maxIterations = 10000)
}
\arguments{
\item{beta}{contact rate.}

\item{dur_lag}{length of time in incubation state.}

\item{r}{recovery rate.}

\item{I_init}{initial number of infectious individuals.}

\item{R_init}{initial number of recovered (immune) individuals.}

\item{N}{total number of individuals in population.}

\item{times}{vector of times at which output should be returned.}

\item{maxIterations}{exit if this number of iterations is reached.}
}
\description{
Draw from stochastic SLIR model, where L is an incubation (lag) stage of defined length, using a compromise between a synchronous and an asynchronous algorithm. The basic algorithm is asynchronous (based on Gillespie's algorithm), but values are only stored and returned at discrete time points. The function still exits automatically at a defined maximum number of iterations.
}

