% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smoothCols}
\alias{smoothCols}
\title{smoothCols}
\usage{
smoothCols(x, xmin = min(x, na.rm = T), xmax = max(x, na.rm = T),
  res = 1000, rawCols = NULL)
}
\arguments{
\item{x}{values from which to obtain colours.}

\item{xmin}{minimum range of x.}

\item{xmax}{maximum range of x.}

\item{res}{number of colours in palette.}

\item{rawCols}{colours that make up the palette. Leave as NULL to use default values, taken from tim.colors().}
}
\description{
Read in continuous values between xmin and xmax and return colours associated with these values, taken from a smoothly varying scale.
}

