% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SterlingFirst}
\alias{SterlingFirst}
\title{SterlingFirst}
\usage{
SterlingFirst(n)
}
\description{
Computes Sterling numbers of the first kind. Results are output in log space, and in general this function is robust to overflow, being capable of handling values as large as n=177.
}
