% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zeroPad}
\alias{zeroPad}
\title{zeroPad}
\usage{
zeroPad(number, padding)
}
\description{
Add leading zeros to number. Can handle negative numbers. For more advanced options see help for the formatC function.
}

