% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filledContour2}
\alias{filledContour2}
\title{filledContour2}
\usage{
filledContour2(z, x = NULL, y = NULL, l = 11, col = bobRedBlue2(),
  orientation = 1, zmin = min(z, na.rm = TRUE), zmax = max(z, na.rm =
  TRUE), main = NA, xlab = "x", ylab = "y", xlab_line = 3,
  ylab_line = 3)
}
\description{
Produces pretty alternative to ordinary filled contour.
}
