% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{normal_loglike}
\alias{normal_loglike}
\title{normal_loglike}
\usage{
normal_loglike(x, sd, priormean, priorsd)
}
\description{
Log-probability of data x integrated over normal likelihood with standard deviation sd and normal prior with mean priorsd and standard deviation priorsd.
Text
}

