% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{RM1_stochastic_sync}
\alias{RM1_stochastic_sync}
\title{RM1_stochastic_sync}
\usage{
RM1_stochastic_sync(a = 0.3, p = 0.9, mu = NULL, u = 22, v = 10,
  r = 1/200, b = 1, c = 1, Eh_init = 0, Ih_init = 10, Em_init = 0,
  Im_init = 0, H = 100, M = 100, times = 0:100)
}
\arguments{
\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. mu = -log(p) unless specified.}

\item{u}{intrinsic incubation period. The number of days from infection to infectiousness in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to infectiousness in a mosquito host.}

\item{r}{daily recovery rate.}

\item{b}{probability a human becomes infected after being bitten by an infected mosquito.}

\item{c}{probability a mosquito becomes infected after biting an infected human.}

\item{Eh_init}{initial number of infected but not infectious humans.}

\item{Ih_init}{initial number of infectious humans.}

\item{Em_init}{initial number of infected but not yet infectious mosquitoes.}

\item{Im_init}{initial number of infectious mosquitoes.}

\item{H}{human population size.}

\item{M}{mosquito population size (number of adult female mosquitoes).}

\item{times}{vector of times at which output should be returned.}
}
\description{
Draw from synchronous stochastic Ross-Macdonald model. Return state of the system at known time points. Results of the synchronous method only match up with the asynchronous method when the time step is small relative to the rates that drive the system.
}

